refactor value names + support container.join

+++ CSharp-Core/src/prompto/value/INumber.cs
@@ -7,7 +7,7 @@ namespace prompto.value
 {
     public interface INumber : IValue
     {
-        long IntegerValue { get; }
-        double DecimalValue { get; }
+        long LongValue { get; }
+        double DoubleValue { get; }
     }
 }
